+++ core/src/services/dropbox/core.rs
@@ -28,17 +28,26 @@ use http::header::CONTENT_LENGTH;
 use http::header::CONTENT_TYPE;
 use http::Request;
 use http::Response;
+use http::StatusCode;
 use serde::Deserialize;
 use serde::Serialize;
 use tokio::sync::Mutex;
 
-use crate::raw::build_rooted_abs_path;
 use crate::raw::new_json_deserialize_error;
 use crate::raw::new_json_serialize_error;
 use crate::raw::new_request_build_error;
 use crate::raw::AsyncBody;
 use crate::raw::HttpClient;
 use crate::raw::IncomingAsyncBody;
+use crate::raw::build_rooted_abs_path;
+use crate::raw::BatchedReply;
+use crate::raw::RpBatch;
+use crate::raw::RpDelete;
+use crate::services::dropbox::backend::DropboxDeleteBatchResponse;
+use crate::services::dropbox::backend::DropboxDeleteBatchResponseEntry;
+use crate::services::dropbox::error::parse_error;
+use crate::types::Error;
+use crate::types::ErrorKind;
 use crate::types::Result;
 
 pub struct DropboxCore {
@@ -156,7 +165,7 @@ impl DropboxCore {
         self.client.send(request).await
     }
 
-    pub async fn dropbox_delete_batch_check(
+    pub async fn dropbox_delete_batch_check_request(
         &self,
         async_job_id: String,
     ) -> Result<Response<IncomingAsyncBody>> {
@@ -173,6 +182,39 @@ impl DropboxCore {
 
         self.sign(&mut request).await?;
         self.client.send(request).await
+    }
+
+    pub async fn dropbox_delete_batch_check(&self, job_id: String) -> Result<RpBatch> {
+        let resp = self.dropbox_delete_batch_check_request(job_id.clone()).await?;
+        let status = resp.status();
+        match status {
+            StatusCode::OK => {
+                let bs = resp.into_body().bytes().await?;
+
+                let decoded_response = serde_json::from_slice::<DropboxDeleteBatchResponse>(&bs)
+                    .map_err(new_json_deserialize_error)?;
+                match decoded_response.tag.as_str() {
+                    "in_progress" => Err(Error::new(
+                        ErrorKind::Unexpected,
+                        "delete batch job still in progress",
+                    )
+                    .set_temporary()),
+                    "complete" => {
+                        let entries = decoded_response.entries.unwrap_or_default();
+                        let results = self.handle_batch_delete_complete_result(entries);
+                        Ok(RpBatch::new(results))
+                    }
+                    _ => Err(Error::new(
+                        ErrorKind::Unexpected,
+                        &format!(
+                            "delete batch check failed with unexpected tag {}",
+                            decoded_response.tag
+                        ),
+                    )),
+                }
+            }
+            _ => Err(parse_error(resp).await?),
+        }
     }
 
     pub async fn dropbox_create_folder(&self, path: &str) -> Result<Response<IncomingAsyncBody>> {
@@ -259,6 +301,43 @@ impl DropboxCore {
         req.headers_mut().insert(header::AUTHORIZATION, value);
 
         Ok(())
+    }
+
+    pub fn handle_batch_delete_complete_result(
+        &self,
+        entries: Vec<DropboxDeleteBatchResponseEntry>,
+    ) -> Vec<(String, Result<BatchedReply>)> {
+        let mut results = Vec::with_capacity(entries.len());
+        for entry in entries {
+            let result = match entry.tag.as_str() {
+                // Only success response has metadata and then path,
+                // so we cannot tell which path failed.
+                "success" => {
+                    let path = entry
+                        .metadata
+                        .expect("metadata should be present")
+                        .path_display;
+                    (path, Ok(RpDelete::default().into()))
+                }
+                "failure" => {
+                    let error = entry.error.expect("error should be present");
+                    let err = Error::new(
+                        ErrorKind::Unexpected,
+                        &format!("delete failed with error {}", error.error_summary),
+                    );
+                    ("".to_string(), Err(err))
+                }
+                _ => (
+                    "".to_string(),
+                    Err(Error::new(
+                        ErrorKind::Unexpected,
+                        &format!("delete failed with unexpected tag {}", entry.tag),
+                    )),
+                ),
+            };
+            results.push(result);
+        }
+        results
     }
 }
